#ELLIE-228 : UI 병합 완료, 스태미나 시스템 병합 완료, 탈진 상태 구현
1. UI 병합 완료, 스태미나 시스템 병합 완료, 탈진 상태 구현

+++ Assets/Scripts/Data/ActionData/Player/StaminaData.cs
@@ -21,7 +21,7 @@ namespace Assets.Scripts.Data.ActionData.Player
 
         public float GetPercentage()
         {
-            return (float)CurrentStamina.Value / (float)maxStamina;
+            return CurrentStamina.Value / maxStamina;
         }
     }
 }
